requirements.txt
Django == 3.1.*
psycopg2-binary ~= 2.8.6
factory-boy ~= 3.2.0
django-choices ~= 1.7.1
django-3-jet ~= 1.0.8
python-dotenv ~= 0.17.0
djangorestframework ~= 3.12.4
requests ~= 2.25.1
drf-yasg ~= 1.20.0
celery~=5.0.0
redis ~= 3.5.3
django-redis ~= 4.12.1
gunicorn == 20.0.4
whitenoise == 5.2.0
Pillow ~= 8.2.0
